# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InvoicePro API Documentation",
    "version": "1.0.0",
    "description": "Exhaustive OpenAPI 3.0 specification for all InvoicePro backend API endpoints (Authentication, Customer Management, Inventory & Stock adjustments, Invoice lifecycle workflows, and Admin User management)."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/inventory": {
      "get": {
        "tags": ["Inventory"],
        "summary": "List Inventory Products",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          {
            "name": "is_active",
            "in": "query",
            "schema": { "type": "boolean" }
          },
          {
            "name": "low_stock",
            "in": "query",
            "schema": { "type": "boolean" }
          },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "List of products"
          }
        }
      },
      "post": {
        "tags": ["Inventory"],
        "summary": "Create Product",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Premium Product" },
                  "description": { "type": "string" },
                  "sku": { "type": "string", "example": "PREM-001" },
                  "unit": { "type": "string", "example": "pcs" },
                  "category": { "type": "string" },
                  "price": { "type": "number", "example": 150.0 },
                  "cost": { "type": "number" },
                  "tax_rate": { "type": "number" },
                  "stock": { "type": "integer" },
                  "reorder_level": { "type": "integer" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          }
        }
      }
    },
    "/api/inventory/{id}": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Get Product by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Product details" }
        }
      },
      "put": {
        "tags": ["Inventory"],
        "summary": "Update Product details",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "price": { "type": "number" },
                  "stock": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Product updated" }
        }
      },
      "delete": {
        "tags": ["Inventory"],
        "summary": "Delete or Deactivate Product",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "hard",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "Set true to permanently delete from DB"
          }
        ],
        "responses": {
          "200": { "description": "Deactivated or deleted successfully" }
        }
      }
    },
    "/api/inventory/{id}/stock-adjust": {
      "post": {
        "tags": ["Inventory"],
        "summary": "Adjust Product Stock",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["in", "out", "set"],
                    "example": "in"
                  },
                  "quantity": { "type": "number", "example": 10 },
                  "notes": { "type": "string" }
                },
                "required": ["type", "quantity"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Stock adjusted successfully" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}